% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected Z-test.R
\name{cor_Z}
\alias{cor_Z}
\title{corrected Z-test of Looney and Jones}
\usage{
cor_Z(x, y, alternative = "two.sided", mu = 0)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values. For tumor data.}

\item{y}{a(non-empty) numeric vector of data values. For normal data}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or"less". You can specify just the initial letter.}

\item{mu}{a number indicating the true value of the mean (or difference in means if you are performing a two sample test).}
}
\value{
A list containing the following components

\code{statistic} the value of the corrected Z-test statistic.

\code{p.value} the p-value for the test.
}
\description{
corrected Z-test of Looney and Jones
}
\examples{
x <- c(3,2,5,13,54,63,22,11,NA,NA,NA)
y <- c(8,3,10,NA,NA,NA,NA,NA,43,96,14)
cor_Z(x,y) # z.stat = -0.4384378  p.value = 0.6610690
}
