% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrZtest.R
\name{corrZtest}
\alias{corrZtest}
\title{Corrected Z-test of Looney and Jones}
\usage{
corrZtest(data, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{data}{Data Frame}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\description{
Corrected Z-test of Looney and Jones
}
\examples{
corrZtest(x)
corrZtest(x, "less")
}
\keyword{stats}
