% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrep.R
\name{dataPrep}
\alias{dataPrep}
\title{Data preparation Process}
\usage{
dataPrep(data)
}
\arguments{
\item{data}{Data Frame}
}
\value{
Common information for further Call
}
\description{
This function read data frame and return common information for further calculation.
Only Designed for internal use within the package.
}
\examples{
dataPrep(x)
}
