% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hete MLE.R
\name{heteMLE}
\alias{heteMLE}
\title{MLE based test of Lin and Stivers under heteroscedasticity}
\usage{
heteMLE(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values. For tumor data.}

\item{y}{a(non-empty) numeric vector of data values. For normal data}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or"less". You can specify just the initial letter.}
}
\value{
A list containing the following components

\code{statistic} the value of the corrected Z-test statistic.

\code{p.value} the p-value for the test.
}
\description{
MLE based test of Lin and Stivers under heteroscedasticity
}
\examples{
heteMLE(x,y,alternative="greater")
}
