% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homo MLE.R
\name{homoMLE}
\alias{homoMLE}
\title{MLE based test of Ekbohm under homoscedasticity}
\usage{
homoMLE(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{vector value of "tumor"}

\item{y}{vector value of "normal"}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or"less". You can specify just the initial letter.}
}
\value{
A list containing the following components

\code{statistic} the value of the corrected Z-test statistic.

\code{p.value} the p-value for the test.
}
\description{
MLE based test of Ekbohm under homoscedasticity
}
\examples{
homeMLE(x,y,alternative = "two.sided")
}
