% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{MLE method procedure under homo- or heteroscedasticity}
\usage{
mle(data, var.equal = TRUE, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{data}{Data Frame}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal. If TRUE then the process homo- MLE otherwise heteroscdasticity case is used.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\description{
MLE method procedure under homo- or heteroscedasticity
}
\examples{
mle(x)
mle(x, var.equal = F, alternative = less)
}
