% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTstat.R
\name{modTstat}
\alias{modTstat}
\title{modified t-statistic of Kim et al.}
\usage{
modTstat(data, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{data}{Data Frame}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
A list with statistics and p value \code{statis}
}
\description{
modified t-statistic of Kim et al.
}
\examples{
modTstat(x, "less")
}
