% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified t-test.R
\name{modttest}
\alias{modttest}
\title{Modified t-test}
\usage{
modttest(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{value of the "tumor"}

\item{y}{value of the "normal"}

\item{alternative}{a character string indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less".}
}
\value{
A list containing the following components:

\code{statistic} the value of weighted z test combination statistic

\code{p.value} the p-value for the test
}
\description{
Modified t-test
}
\examples{
x <- c(1,2,3,4,NA,NA,1)
y <- c(1,2,5)
my.mod.t.test(x,y)
}
