% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\title{Output well-organized list for test}
\usage{
output(statistic = NULL, pv, METHOD)
}
\arguments{
\item{statistic}{The value of the Z-statistic}

\item{pv}{The p value for the test}

\item{METHOD}{a character string indicating what type of Z-test was performed}
}
\value{

}
\description{
Only Designed for internal use within the package.
}
