% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{pval}
\alias{pval}
\title{P value for Z-test
Only Designed for internal use within the package.}
\usage{
pval(statistic, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{statistic}{The value of the Z-statistic}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
p value
}
\description{
P value for Z-test
Only Designed for internal use within the package.
}
