% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted z.R
\name{weighted.z.test}
\alias{weighted.z.test}
\title{Weighted Z test combination}
\usage{
weighted.z.test(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numberic vector of data values.}

\item{y}{a (non-empty) numberic vector of data values.}

\item{alternative}{a character string indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less".}
}
\value{
A list containing the following components:
\item{statistic}{the value of weighted z test combination statistic}
\item{p.value}{the p-value for the test}
\item{METHOD}{the method we implement}
}
\description{
Weighted Z test combination
}
\examples{
x <- c(3,5,8,12,9,23,42,NA,58,NA,NA)
y <- c(10,5,NA,8,NA,NA,18,NA,45,66,44)
weighted.z.test(x,y,alternative="two.sided") #pc=1.234111  p.value=0.2171614
}
