% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightZ.R
\name{weightZ}
\alias{weightZ}
\title{Weighted Z-test Combination}
\usage{
weightZ(data, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{data}{Data Frame}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
A list with statistics and p value

\code{statistic} Test statistic

\code{p.value} P value of test

\code{METHOD} What method has been implemented
}
\description{
Weighted Z-test Combination
}
\examples{
weightZ(x, less)
}
